Added a radical line of code to the end of the Npolapp.cc (Main) program.  std::quick_exit(0)  This line of code stops the program in its tracks and forces the operating systems to sweep up the mess.  I'm using this to end a crash that occurs when void main() exits.
Basically, when I use the biasing histogram method and the program ends ... literially when void main() returns it crashes when it calls some G4 code to complete the clean up process.  Daniel thinks it might be a G4 bug, however, I can't find anyone else with the same issue and formulating a bug report isn't something I've done before.  I have checked my code for possible clean up errors, made sure the code aligned with examples, and all around dumped the errors to the screen.  Nothing worked or appeared out of line with the specified examples and practices.  On my local machines this isn't a problem as the code just spits out errors and then moves onto the next level of the submission script.  However, on JLAB there is a memory limit and many times the error pusshes the system over its memory allocation limit and the run is ended.  This allows for JLAB runs to continue.

  Anyway, I works now even though it is very curious.  It doesn't seem to cause other issues.