(1) Added a line to push the particle string name to the vertexMap filled during a call to fillVertexMap in NpolPhysicsVariables.
(2) Added a requirement to fillVertexMap that requires that the vertices have to be in the same volume from which the event occured. This will eliminate the issue with vertices being included from multiple scattering events.

(3) Added a printVertexMap method to NpolPhysicsVariables so the user can scan the events which pass their requirements for the vertices and check for patterns and fun stuff.