Added a method to NpolPhysicsVariables Class to return the initial neutron's energy from the vertexMap.
Found a fundamental error in determining the "leading" particle from inelastic/quasielastic scatterings.  In these, the initial neutron is 'stopped' and a new neutron (based on final state stuff) is generated which will have a momentum close to what the initial neutron had (or so ... it depends on the exact details).  Either way, in most of the inelastic and quasielastic cases we were returning the highest momentum neutron's TID which is the neutron that is scattered.  No wonder the momenta didn't look right.  Now, the method finds this neutron and then re-runs the map search excluding this track from the search.

Lastly, in computeElasticMomentum, we now send it the neutron's energy as well as it's momentum.  Not really necessary but that's the way it is currently written.